Refactoring du code exportManager Renommage de la class DiffStatsManager en DiffComputer
